requirements.txt
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.6.15
httpx==0.24.1
idna==3.10
nest-asyncio==1.6.0
python-telegram-bot==20.3
sniffio==1.3.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
